TG-21 - model implementation improvements

+++ app/src/main/java/model/Mark.java
@@ -45,7 +45,7 @@ public class Mark implements MarkIf
     }
 
     @Override
-    public double getMarkValue()
+    public float getMarkValue()
     {
         return markValue.getMarkValue();
     }
