+++ app/src/main/java/model/MarkIf.java
@@ -14,5 +14,5 @@ public interface MarkIf
 
     int getUniversitySubjectId();
 
-    double getMarkValue();
+    float getMarkValue();
 }
